[arcane,wrapperc#] Déplace 'ItemIndexArrayView' et 'ItemVectorView' dans son propre fichier.
